requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pywebview>=4.4.1
pymupdf>=1.23.0
ebooklib>=0.18
beautifulsoup4>=4.12.0
requests>=2.31.0
Pillow>=10.0.0
python-multipart>=0.0.9
